[proliferate] split organization settings pane
